Add shocks
GDP figure is still not working properly.
I try to add the shocks to the code: until now shocks were not added,
the model only simulated a convergence to steady state starting from a
different point (actually given nothing is specified as endval the model
goes back to initval in the end of the simulation, as it is a two
boundary problem with the same initial and end value. In the middle it
stays on the steady state)